requirements.txt
speechrecognition
pyaudio
faster-whisper
openai-whisper
pydub
